# Change an array to a tuple to improve parsing performance by 10x