@iamed2 2021-07-24
During experimentation, I learned that `RoundingMode[RoundThrows, RoundNearest, RoundToZero]` is only ~10% slower than `(RoundThrows, RoundNearest, RoundToZero)`. I'm going to see if I can make a Base improvement that achieves the speedup of `RoundingMode[RoundThrows, RoundNearest, RoundToZero]` for `[RoundThrows, RoundNearest, RoundToZero]` automatically.